I hereby agree to the terms of the CLA available at: https://databend.rs/policies/cla/

## Summary

Summary about this PR



Before:
```
:) SELECT sum(  number  )  FROM numbers_mt(1000000000)  where true;
+--------------------+
| sum(number)        |
+--------------------+
| 499999999500000000 |
+--------------------+
1 row in set (0.18 sec)
Read 1000000000 rows, 8 GB in 0.171 sec., 5.84 billion rows/sec., 46.69 GB/sec.
```

After:
```

:) SELECT sum(  number  )  FROM numbers_mt(1000000000) where true  ;
+--------------------+
| sum(number)        |
+--------------------+
| 499999999500000000 |
+--------------------+
1 row in set (0.07 sec)
Read 1000000000 rows, 8 GB in 0.070 sec., 14.2 billion rows/sec., 113.57 GB/sec.


```

## Changelog

- Improvement
- Performance Improvement


## Related Issues

Fixes #issue

## Test Plan

Unit Tests

Stateless Tests